third_party/blink/web_tests/http/tests/websocket/compressed-control-frame-expected.txt

Test whether a compressed control frame is rejected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
